backend/requirements.txt
Flask
flask_cors
datetime
mysql.connector
flask_limiter
Flask-SQLAlchemy
PyMySQL
